pyproject.toml
[project]
name = "writeups"
version = "0.1.0"
description = "stupid writeups site"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "markdown>=3.7",
    "pygments>=2.19.2",
    "pymdown-extensions>=10.18",
    "python-frontmatter>=1.1.0",
    "python-markdown-math>=0.9",
    "staticjinja>=5.0.0",
    "watchdog>=6.0.0",
]
